refactor(agent): relocate agent-run.ts → src/agent/run-adapter.ts
Final step of the agent-layer relocation: the chat-side adapter that
turns AgentEvent streams into {responseText, errorMessage} now lives
under src/agent/ alongside init.ts and system-prompt.ts.
