+++ src/agent/run-adapter.test.ts
@@ -1,11 +1,11 @@
 // Direct unit coverage for consumeRootRun + cancelRunBySessionId.
 
 import { describe, it, expect, vi } from "vitest";
-import { consumeRootRun, cancelRunBySessionId } from "./agent-run.js";
-import { AgentRuntime } from "../agents/runtime.js";
-import type { RegisteredAgent, SendMessageRequest } from "../agents/types.js";
+import { consumeRootRun, cancelRunBySessionId } from "./run-adapter.js";
+import { AgentRuntime } from "../legacy/agents/runtime.js";
+import type { RegisteredAgent, SendMessageRequest } from "../legacy/agents/types.js";
 import type { AgentEvent } from "@mariozechner/pi-agent-core";
-import { createLogger } from "../../logging/logger.js";
+import { createLogger } from "../logging/logger.js";
 
 const log = createLogger("test:agent-run");
